requirements.txt
crewai[tools]>=0.114.0,<1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain>=0.0.325
langchain-community>=0.0.1
langchain-google-genai>=0.0.1